Move pvalues into separate file
